refactor(LC042): split missing-query-tags analysis helpers
- separate query-chain analysis from analyzer orchestration
- keep analyzer behavior unchanged while reducing local complexity
